+++ src/adapter/src/coord/ddl.rs
@@ -70,7 +70,7 @@ impl<S: Append + 'static> Coordinator<S> {
     /// function successfully returns on any built `DataflowDesc`.
     ///
     /// [`CatalogState`]: crate::catalog::CatalogState
-    /// [`DataflowDesc`]: mz_compute_client::command::DataflowDesc
+    /// [`DataflowDesc`]: mz_compute_client::types::dataflows::DataflowDesc
     #[tracing::instrument(level = "debug", skip_all)]
     pub(crate) async fn catalog_transact_with<F, R>(
         &mut self,
