+++ src/compute-client/src/command.rs
@@ -9,7 +9,7 @@
 
 //! Compute layer commands.
 
-use std::collections::{BTreeMap, BTreeSet};
+use std::collections::BTreeSet;
 use std::num::NonZeroI64;
 
 use proptest::prelude::{any, Arbitrary};
@@ -20,21 +20,15 @@ use serde::{Deserialize, Serialize};
 use timely::progress::frontier::Antichain;
 use uuid::Uuid;
 
-use mz_expr::{
-    CollectionPlan, MirRelationExpr, MirScalarExpr, OptimizedMirRelationExpr, RowSetFinishing,
-};
+use mz_expr::RowSetFinishing;
 use mz_ore::tracing::OpenTelemetryContext;
-use mz_proto::{any_uuid, IntoRustIfSome, ProtoMapEntry, ProtoType, RustType, TryFromProtoError};
-use mz_repr::{GlobalId, RelationType, Row};
+use mz_proto::{any_uuid, IntoRustIfSome, ProtoType, RustType, TryFromProtoError};
+use mz_repr::{GlobalId, Row};
 use mz_storage_client::client::ProtoAllowCompaction;
 use mz_storage_client::controller::CollectionMetadata;
 
-use crate::command::proto_dataflow_description::{
-    ProtoIndexExport, ProtoIndexImport, ProtoSinkExport, ProtoSourceImport,
-};
 use crate::logging::LoggingConfig;
-use crate::plan::Plan;
-use crate::sinks::{ComputeSinkConnection, ComputeSinkDesc};
+use crate::types::dataflows::DataflowDescription;
 
 include!(concat!(env!("OUT_DIR"), "/mz_compute_client.command.rs"));
 
@@ -359,590 +353,6 @@ impl RustType<ProtoCommunicationConfig> for CommunicationConfig {
             process: proto.process.into_rust()?,
             workers: proto.workers.into_rust()?,
             addresses: proto.addresses.into_rust()?,
-        })
-    }
-}
-
-/// A commonly used name for dataflows contain MIR expressions.
-pub type DataflowDesc = DataflowDescription<OptimizedMirRelationExpr, ()>;
-
-/// An association of a global identifier to an expression.
-#[derive(Arbitrary, Clone, Debug, Eq, PartialEq, Serialize, Deserialize)]
-pub struct BuildDesc<P> {
-    pub id: GlobalId,
-    pub plan: P,
-}
-
-impl RustType<ProtoBuildDesc> for BuildDesc<crate::plan::Plan> {
-    fn into_proto(&self) -> ProtoBuildDesc {
-        ProtoBuildDesc {
-            id: Some(self.id.into_proto()),
-            plan: Some(self.plan.into_proto()),
-        }
-    }
-
-    fn from_proto(x: ProtoBuildDesc) -> Result<Self, TryFromProtoError> {
-        Ok(BuildDesc {
-            id: x.id.into_rust_if_some("ProtoBuildDesc::id")?,
-            plan: x.plan.into_rust_if_some("ProtoBuildDesc::plan")?,
-        })
-    }
-}
-
-/// A description of an instantiation of a source.
-///
-/// This includes a description of the source, but additionally any
-/// context-dependent options like the ability to apply filtering and
-/// projection to the records as they emerge.
-#[derive(Arbitrary, Debug, Clone, Eq, PartialEq, Serialize, Deserialize)]
-pub struct SourceInstanceDesc<M> {
-    /// Arguments for this instantiation of the source.
-    pub arguments: SourceInstanceArguments,
-    /// Additional metadata used by the storage client of a compute instance to read it.
-    pub storage_metadata: M,
-    /// The relation type of this source
-    pub typ: RelationType,
-}
-
-impl RustType<ProtoSourceInstanceDesc> for SourceInstanceDesc<CollectionMetadata> {
-    fn into_proto(&self) -> ProtoSourceInstanceDesc {
-        ProtoSourceInstanceDesc {
-            arguments: Some(self.arguments.into_proto()),
-            storage_metadata: Some(self.storage_metadata.into_proto()),
-            typ: Some(self.typ.into_proto()),
-        }
-    }
-
-    fn from_proto(proto: ProtoSourceInstanceDesc) -> Result<Self, TryFromProtoError> {
-        Ok(SourceInstanceDesc {
-            arguments: proto
-                .arguments
-                .into_rust_if_some("ProtoSourceInstanceDesc::arguments")?,
-            storage_metadata: proto
-                .storage_metadata
-                .into_rust_if_some("ProtoSourceInstanceDesc::storage_metadata")?,
-            typ: proto
-                .typ
-                .into_rust_if_some("ProtoSourceInstanceDesc::typ")?,
-        })
-    }
-}
-
-/// Per-source construction arguments.
-#[derive(Arbitrary, Debug, Clone, Eq, PartialEq, Serialize, Deserialize)]
-pub struct SourceInstanceArguments {
-    /// Linear operators to be applied record-by-record.
-    pub operators: Option<mz_expr::MapFilterProject>,
-}
-
-impl RustType<ProtoSourceInstanceArguments> for SourceInstanceArguments {
-    fn into_proto(&self) -> ProtoSourceInstanceArguments {
-        ProtoSourceInstanceArguments {
-            operators: self.operators.into_proto(),
-        }
-    }
-
-    fn from_proto(proto: ProtoSourceInstanceArguments) -> Result<Self, TryFromProtoError> {
-        Ok(SourceInstanceArguments {
-            operators: proto.operators.into_rust()?,
-        })
-    }
-}
-
-/// Type alias for source subscriptions, (dataflow_id, source_id).
-pub type SourceInstanceId = (uuid::Uuid, mz_repr::GlobalId);
-
-/// A description of a dataflow to construct and results to surface.
-#[derive(Clone, Debug, Serialize, Deserialize, Eq, PartialEq)]
-pub struct DataflowDescription<P, S = (), T = mz_repr::Timestamp> {
-    /// Sources instantiations made available to the dataflow pair with monotonicity information.
-    pub source_imports: BTreeMap<GlobalId, (SourceInstanceDesc<S>, bool)>,
-    /// Indexes made available to the dataflow.
-    /// (id of new index, description of index, relationtype of base source/view, monotonic)
-    pub index_imports: BTreeMap<GlobalId, (IndexDesc, RelationType, bool)>,
-    /// Views and indexes to be built and stored in the local context.
-    /// Objects must be built in the specific order, as there may be
-    /// dependencies of later objects on prior identifiers.
-    pub objects_to_build: Vec<BuildDesc<P>>,
-    /// Indexes to be made available to be shared with other dataflows
-    /// (id of new index, description of index, relationtype of base source/view)
-    pub index_exports: BTreeMap<GlobalId, (IndexDesc, RelationType)>,
-    /// sinks to be created
-    /// (id of new sink, description of sink)
-    pub sink_exports: BTreeMap<GlobalId, ComputeSinkDesc<S, T>>,
-    /// An optional frontier to which inputs should be advanced.
-    ///
-    /// If this is set, it should override the default setting determined by
-    /// the upper bound of `since` frontiers contributing to the dataflow.
-    /// It is an error for this to be set to a frontier not beyond that default.
-    pub as_of: Option<Antichain<T>>,
-    /// Frontier beyond which the dataflow should not execute.
-    /// Specifically, updates at times greater or equal to this frontier are suppressed.
-    /// This is often set to `as_of + 1` to enable "batch" computations.
-    pub until: Antichain<T>,
-    /// Human readable name
-    pub debug_name: String,
-}
-
-fn any_source_import(
-) -> impl Strategy<Value = (GlobalId, (SourceInstanceDesc<CollectionMetadata>, bool))> {
-    (
-        any::<GlobalId>(),
-        any::<(SourceInstanceDesc<CollectionMetadata>, bool)>(),
-    )
-}
-
-proptest::prop_compose! {
-    fn any_dataflow_index_import()(
-        id in any::<GlobalId>(),
-        index in any::<IndexDesc>(),
-        typ in any::<RelationType>(),
-        monotonic in any::<bool>(),
-    ) -> (GlobalId, (IndexDesc, RelationType, bool)) {
-        (id, (index, typ, monotonic))
-    }
-}
-
-proptest::prop_compose! {
-    fn any_dataflow_index_export()(
-        id in any::<GlobalId>(),
-        index in any::<IndexDesc>(),
-        typ in any::<RelationType>(),
-    ) -> (GlobalId, (IndexDesc, RelationType)) {
-        (id, (index, typ))
-    }
-}
-
-proptest::prop_compose! {
-    fn any_dataflow_description()(
-        source_imports in proptest::collection::vec(any_source_import(), 1..3),
-        index_imports in proptest::collection::vec(any_dataflow_index_import(), 1..3),
-        objects_to_build in proptest::collection::vec(any::<BuildDesc<Plan>>(), 1..3),
-        index_exports in proptest::collection::vec(any_dataflow_index_export(), 1..3),
-        sink_descs in proptest::collection::vec(
-            any::<(GlobalId, ComputeSinkDesc<CollectionMetadata, mz_repr::Timestamp>)>(),
-            1..3,
-        ),
-        as_of_some in any::<bool>(),
-        as_of in proptest::collection::vec(any::<mz_repr::Timestamp>(), 1..5),
-        debug_name in ".*",
-    ) -> DataflowDescription<Plan, CollectionMetadata, mz_repr::Timestamp> {
-        DataflowDescription {
-            source_imports: BTreeMap::from_iter(source_imports.into_iter()),
-            index_imports: BTreeMap::from_iter(index_imports.into_iter()),
-            objects_to_build,
-            index_exports: BTreeMap::from_iter(index_exports.into_iter()),
-            sink_exports: BTreeMap::from_iter(
-                sink_descs.into_iter(),
-            ),
-            as_of: if as_of_some {
-                Some(Antichain::from(as_of))
-            } else {
-                None
-            },
-            until: Antichain::new(),
-            debug_name,
-        }
-    }
-}
-
-impl Arbitrary for DataflowDescription<Plan, CollectionMetadata, mz_repr::Timestamp> {
-    type Strategy = BoxedStrategy<Self>;
-    type Parameters = ();
-
-    fn arbitrary_with(_: Self::Parameters) -> Self::Strategy {
-        any_dataflow_description().boxed()
-    }
-}
-
-impl<T> DataflowDescription<OptimizedMirRelationExpr, (), T> {
-    /// Creates a new dataflow description with a human-readable name.
-    pub fn new(name: String) -> Self {
-        Self {
-            source_imports: Default::default(),
-            index_imports: Default::default(),
-            objects_to_build: Vec::new(),
-            index_exports: Default::default(),
-            sink_exports: Default::default(),
-            as_of: Default::default(),
-            until: Antichain::new(),
-            debug_name: name,
-        }
-    }
-
-    /// Imports a previously exported index.
-    ///
-    /// This method makes available an index previously exported as `id`, identified
-    /// to the query by `description` (which names the view the index arranges, and
-    /// the keys by which it is arranged).
-    pub fn import_index(
-        &mut self,
-        id: GlobalId,
-        description: IndexDesc,
-        typ: RelationType,
-        monotonic: bool,
-    ) {
-        self.index_imports.insert(id, (description, typ, monotonic));
-    }
-
-    /// Imports a source and makes it available as `id`.
-    pub fn import_source(&mut self, id: GlobalId, typ: RelationType, monotonic: bool) {
-        // Import the source with no linear operators applied to it.
-        // They may be populated by whole-dataflow optimization.
-        self.source_imports.insert(
-            id,
-            (
-                SourceInstanceDesc {
-                    storage_metadata: (),
-                    arguments: SourceInstanceArguments { operators: None },
-                    typ,
-                },
-                monotonic,
-            ),
-        );
-    }
-
-    /// Binds to `id` the relation expression `plan`.
-    pub fn insert_plan(&mut self, id: GlobalId, plan: OptimizedMirRelationExpr) {
-        self.objects_to_build.push(BuildDesc { id, plan });
-    }
-
-    /// Exports as `id` an index described by `description`.
-    ///
-    /// Future uses of `import_index` in other dataflow descriptions may use `id`,
-    /// as long as this dataflow has not been terminated in the meantime.
-    pub fn export_index(&mut self, id: GlobalId, description: IndexDesc, on_type: RelationType) {
-        // We first create a "view" named `id` that ensures that the
-        // data are correctly arranged and available for export.
-        self.insert_plan(
-            id,
-            OptimizedMirRelationExpr::declare_optimized(MirRelationExpr::ArrangeBy {
-                input: Box::new(MirRelationExpr::global_get(
-                    description.on_id,
-                    on_type.clone(),
-                )),
-                keys: vec![description.key.clone()],
-            }),
-        );
-        self.index_exports.insert(id, (description, on_type));
-    }
-
-    /// Exports as `id` a sink described by `description`.
-    pub fn export_sink(&mut self, id: GlobalId, description: ComputeSinkDesc<(), T>) {
-        self.sink_exports.insert(id, description);
-    }
-
-    /// Returns true iff `id` is already imported.
-    pub fn is_imported(&self, id: &GlobalId) -> bool {
-        self.objects_to_build.iter().any(|bd| &bd.id == id)
-            || self.source_imports.keys().any(|i| i == id)
-    }
-
-    /// Assigns the `as_of` frontier to the supplied argument.
-    ///
-    /// This method allows the dataflow to indicate a frontier up through
-    /// which all times should be advanced. This can be done for at least
-    /// two reasons: 1. correctness and 2. performance.
-    ///
-    /// Correctness may require an `as_of` to ensure that historical detail
-    /// is consolidated at representative times that do not present specific
-    /// detail that is not specifically correct. For example, updates may be
-    /// compacted to times that are no longer the source times, but instead
-    /// some byproduct of when compaction was executed; we should not present
-    /// those specific times as meaningfully different from other equivalent
-    /// times.
-    ///
-    /// Performance may benefit from an aggressive `as_of` as it reduces the
-    /// number of distinct moments at which collections vary. Differential
-    /// dataflow will refresh its outputs at each time its inputs change and
-    /// to moderate that we can minimize the volume of distinct input times
-    /// as much as possible.
-    ///
-    /// Generally, one should consider setting `as_of` at least to the `since`
-    /// frontiers of contributing data sources and as aggressively as the
-    /// computation permits.
-    pub fn set_as_of(&mut self, as_of: Antichain<T>) {
-        self.as_of = Some(as_of);
-    }
-
-    /// The number of columns associated with an identifier in the dataflow.
-    pub fn arity_of(&self, id: &GlobalId) -> usize {
-        for (source_id, (source, _monotonic)) in self.source_imports.iter() {
-            if source_id == id {
-                return source.typ.arity();
-            }
-        }
-        for (desc, typ, _monotonic) in self.index_imports.values() {
-            if &desc.on_id == id {
-                return typ.arity();
-            }
-        }
-        for desc in self.objects_to_build.iter() {
-            if &desc.id == id {
-                return desc.plan.arity();
-            }
-        }
-        panic!("GlobalId {} not found in DataflowDesc", id);
-    }
-}
-
-impl<P, S, T> DataflowDescription<P, S, T>
-where
-    P: CollectionPlan,
-{
-    /// Identifiers of exported objects (indexes and sinks).
-    pub fn export_ids(&self) -> impl Iterator<Item = GlobalId> + '_ {
-        self.index_exports
-            .keys()
-            .chain(self.sink_exports.keys())
-            .cloned()
-    }
-
-    /// Identifiers of exported subscribe sinks.
-    pub fn subscribe_ids(&self) -> impl Iterator<Item = GlobalId> + '_ {
-        self.sink_exports
-            .iter()
-            .filter_map(|(id, desc)| match desc.connection {
-                ComputeSinkConnection::Subscribe(_) => Some(*id),
-                _ => None,
-            })
-    }
-
-    /// Returns the description of the object to build with the specified
-    /// identifier.
-    ///
-    /// # Panics
-    ///
-    /// Panics if `id` is not present in `objects_to_build` exactly once.
-    pub fn build_desc(&self, id: GlobalId) -> &BuildDesc<P> {
-        let mut builds = self.objects_to_build.iter().filter(|build| build.id == id);
-        let build = builds
-            .next()
-            .unwrap_or_else(|| panic!("object to build id {id} unexpectedly missing"));
-        assert!(builds.next().is_none());
-        build
-    }
-
-    /// Computes the set of identifiers upon which the specified collection
-    /// identifier depends.
-    ///
-    /// `id` must specify a valid object in `objects_to_build`.
-    pub fn depends_on(&self, collection_id: GlobalId) -> BTreeSet<GlobalId> {
-        let mut out = BTreeSet::new();
-        self.depends_on_into(collection_id, &mut out);
-        out
-    }
-
-    /// Like `depends_on`, but appends to an existing `BTreeSet`.
-    pub fn depends_on_into(&self, collection_id: GlobalId, out: &mut BTreeSet<GlobalId>) {
-        if self.source_imports.contains_key(&collection_id) {
-            // The collection is provided by an imported source. Report the
-            // dependency on the source.
-            out.insert(collection_id);
-            return;
-        }
-
-        // NOTE(benesch): we're not smart enough here to know *which* index
-        // for the collection will be used, if one exists, so we have to report
-        // the dependency on all of them.
-        let mut found_index = false;
-        for (index_id, (desc, _typ, _monotonic)) in &self.index_imports {
-            if desc.on_id == collection_id {
-                // The collection is provided by an imported index. Report the
-                // dependency on the index.
-                out.insert(*index_id);
-                found_index = true;
-            }
-        }
-        if found_index {
-            return;
-        }
-
-        // The collection is not provided by a source or imported index.
-        // It must be a collection whose plan we have handy. Recurse.
-        let build = self.build_desc(collection_id);
-        for id in build.plan.depends_on() {
-            self.depends_on_into(id, out)
-        }
-    }
-}
-
-impl<P: PartialEq, S: PartialEq, T: timely::PartialOrder> DataflowDescription<P, S, T> {
-    /// Determine if a dataflow description is compatible with this dataflow description.
-    ///
-    /// Compatible dataflows have equal exports, imports, and objects to build. The `as_of` of
-    /// the receiver has to be less equal the `other` `as_of`.
-    ///
-    // TODO: The semantics of this function are only useful for command reconciliation at the moment.
-    pub fn compatible_with(&self, other: &Self) -> bool {
-        let equality = self.index_exports == other.index_exports
-            && self.sink_exports == other.sink_exports
-            && self.objects_to_build == other.objects_to_build
-            && self.index_imports == other.index_imports
-            && self.source_imports == other.source_imports;
-        let partial = if let (Some(as_of), Some(other_as_of)) = (&self.as_of, &other.as_of) {
-            timely::PartialOrder::less_equal(as_of, other_as_of)
-        } else {
-            false
-        };
-        equality && partial
-    }
-}
-
-impl RustType<ProtoDataflowDescription>
-    for DataflowDescription<crate::plan::Plan, CollectionMetadata>
-{
-    fn into_proto(&self) -> ProtoDataflowDescription {
-        ProtoDataflowDescription {
-            source_imports: self.source_imports.into_proto(),
-            index_imports: self.index_imports.into_proto(),
-            objects_to_build: self.objects_to_build.into_proto(),
-            index_exports: self.index_exports.into_proto(),
-            sink_exports: self.sink_exports.into_proto(),
-            as_of: self.as_of.into_proto(),
-            until: Some(self.until.into_proto()),
-            debug_name: self.debug_name.clone(),
-        }
-    }
-
-    fn from_proto(proto: ProtoDataflowDescription) -> Result<Self, TryFromProtoError> {
-        Ok(DataflowDescription {
-            source_imports: proto.source_imports.into_rust()?,
-            index_imports: proto.index_imports.into_rust()?,
-            objects_to_build: proto.objects_to_build.into_rust()?,
-            index_exports: proto.index_exports.into_rust()?,
-            sink_exports: proto.sink_exports.into_rust()?,
-            as_of: proto.as_of.map(|x| x.into_rust()).transpose()?,
-            until: proto
-                .until
-                .map(|x| x.into_rust())
-                .transpose()?
-                .unwrap_or_else(Antichain::new),
-            debug_name: proto.debug_name,
-        })
-    }
-}
-
-impl ProtoMapEntry<GlobalId, (SourceInstanceDesc<CollectionMetadata>, bool)> for ProtoSourceImport {
-    fn from_rust<'a>(
-        entry: (
-            &'a GlobalId,
-            &'a (SourceInstanceDesc<CollectionMetadata>, bool),
-        ),
-    ) -> Self {
-        ProtoSourceImport {
-            id: Some(entry.0.into_proto()),
-            source_instance_desc: Some(entry.1 .0.into_proto()),
-            monotonic: entry.1 .1.into_proto(),
-        }
-    }
-
-    fn into_rust(
-        self,
-    ) -> Result<(GlobalId, (SourceInstanceDesc<CollectionMetadata>, bool)), TryFromProtoError> {
-        Ok((
-            self.id.into_rust_if_some("ProtoSourceImport::id")?,
-            (
-                self.source_instance_desc
-                    .into_rust_if_some("ProtoSourceImport::source_instance_desc")?,
-                self.monotonic.into_rust()?,
-            ),
-        ))
-    }
-}
-
-impl ProtoMapEntry<GlobalId, (IndexDesc, RelationType, bool)> for ProtoIndexImport {
-    fn from_rust<'a>(
-        (id, (index_desc, typ, monotonic)): (&'a GlobalId, &'a (IndexDesc, RelationType, bool)),
-    ) -> Self {
-        ProtoIndexImport {
-            id: Some(id.into_proto()),
-            index_desc: Some(index_desc.into_proto()),
-            typ: Some(typ.into_proto()),
-            monotonic: monotonic.into_proto(),
-        }
-    }
-
-    fn into_rust(self) -> Result<(GlobalId, (IndexDesc, RelationType, bool)), TryFromProtoError> {
-        Ok((
-            self.id.into_rust_if_some("ProtoIndex::id")?,
-            (
-                self.index_desc
-                    .into_rust_if_some("ProtoIndexImport::index_desc")?,
-                self.typ.into_rust_if_some("ProtoIndexImport::typ")?,
-                self.monotonic.into_rust()?,
-            ),
-        ))
-    }
-}
-
-impl ProtoMapEntry<GlobalId, (IndexDesc, RelationType)> for ProtoIndexExport {
-    fn from_rust<'a>(
-        (id, (index_desc, typ)): (&'a GlobalId, &'a (IndexDesc, RelationType)),
-    ) -> Self {
-        ProtoIndexExport {
-            id: Some(id.into_proto()),
-            index_desc: Some(index_desc.into_proto()),
-            typ: Some(typ.into_proto()),
-        }
-    }
-
-    fn into_rust(self) -> Result<(GlobalId, (IndexDesc, RelationType)), TryFromProtoError> {
-        Ok((
-            self.id.into_rust_if_some("ProtoIndexExport::id")?,
-            (
-                self.index_desc
-                    .into_rust_if_some("ProtoIndexExport::index_desc")?,
-                self.typ.into_rust_if_some("ProtoIndexExport::typ")?,
-            ),
-        ))
-    }
-}
-
-impl ProtoMapEntry<GlobalId, ComputeSinkDesc<CollectionMetadata>> for ProtoSinkExport {
-    fn from_rust<'a>(
-        (id, sink_desc): (&'a GlobalId, &'a ComputeSinkDesc<CollectionMetadata>),
-    ) -> Self {
-        ProtoSinkExport {
-            id: Some(id.into_proto()),
-            sink_desc: Some(sink_desc.into_proto()),
-        }
-    }
-
-    fn into_rust(
-        self,
-    ) -> Result<(GlobalId, ComputeSinkDesc<CollectionMetadata>), TryFromProtoError> {
-        Ok((
-            self.id.into_rust_if_some("ProtoSinkExport::id")?,
-            self.sink_desc
-                .into_rust_if_some("ProtoSinkExport::sink_desc")?,
-        ))
-    }
-}
-
-/// An index storing processed updates so they can be queried
-/// or reused in other computations
-#[derive(Arbitrary, Clone, Debug, Eq, PartialEq, Serialize, Deserialize, Hash)]
-pub struct IndexDesc {
-    /// Identity of the collection the index is on.
-    pub on_id: GlobalId,
-    /// Expressions to be arranged, in order of decreasing primacy.
-    #[proptest(strategy = "proptest::collection::vec(any::<MirScalarExpr>(), 1..3)")]
-    pub key: Vec<MirScalarExpr>,
-}
-
-impl RustType<ProtoIndexDesc> for IndexDesc {
-    fn into_proto(&self) -> ProtoIndexDesc {
-        ProtoIndexDesc {
-            on_id: Some(self.on_id.into_proto()),
-            key: self.key.into_proto(),
-        }
-    }
-
-    fn from_proto(proto: ProtoIndexDesc) -> Result<Self, TryFromProtoError> {
-        Ok(IndexDesc {
-            on_id: proto.on_id.into_rust_if_some("ProtoIndexDesc::on_id")?,
-            key: proto.key.into_rust()?,
         })
     }
 }
@@ -1284,15 +694,6 @@ mod tests {
         #[ignore]
         fn compute_command_protobuf_roundtrip(expect in any::<ComputeCommand<mz_repr::Timestamp>>() ) {
             let actual = protobuf_roundtrip::<_, ProtoComputeCommand>(&expect);
-            assert!(actual.is_ok());
-            assert_eq!(actual.unwrap(), expect);
-        }
-
-        // TODO: Unignore after fixing #14543.
-        #[test]
-        #[ignore]
-        fn dataflow_description_protobuf_roundtrip(expect in any::<DataflowDescription<Plan, CollectionMetadata, mz_repr::Timestamp>>()) {
-            let actual = protobuf_roundtrip::<_, ProtoDataflowDescription>(&expect);
             assert!(actual.is_ok());
             assert_eq!(actual.unwrap(), expect);
         }
