+++ src/compute-client/src/explain.rs
@@ -37,7 +37,7 @@ use mz_ore::str::{bracketed, separated};
 use mz_repr::explain_new::ExprHumanizer;
 use mz_repr::GlobalId;
 
-use crate::command::DataflowDescription;
+use crate::types::dataflows::DataflowDescription;
 
 pub trait ViewFormatter<ViewExpr> {
     fn fmt_source_body(&self, f: &mut fmt::Formatter, operator: &MapFilterProject) -> fmt::Result;
