+++ src/compute-client/src/types/sinks.rs
@@ -7,6 +7,8 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
+//! Types for describing dataflow sinks.
+
 use proptest::prelude::{any, Arbitrary, BoxedStrategy, Strategy};
 use proptest_derive::Arbitrary;
 use serde::{Deserialize, Serialize};
@@ -16,7 +18,10 @@ use mz_proto::{IntoRustIfSome, ProtoType, RustType, TryFromProtoError};
 use mz_repr::{GlobalId, RelationDesc};
 use mz_storage_client::controller::CollectionMetadata;
 
-include!(concat!(env!("OUT_DIR"), "/mz_compute_client.sinks.rs"));
+include!(concat!(
+    env!("OUT_DIR"),
+    "/mz_compute_client.types.sinks.rs"
+));
 
 /// A sink for updates to a relational collection.
 #[derive(Clone, Debug, Serialize, Deserialize, Eq, PartialEq)]
