+++ src/compute-client/src/types/sources.proto
@@ -1,0 +1,26 @@
+// Copyright Materialize, Inc. and contributors. All rights reserved.
+//
+// Use of this software is governed by the Business Source License
+// included in the LICENSE file.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0.
+
+syntax = "proto3";
+
+import "expr/src/linear.proto";
+import "repr/src/relation_and_scalar.proto";
+import "storage-client/src/controller.proto";
+
+package mz_compute_client.types.sources;
+
+message ProtoSourceInstanceDesc {
+    ProtoSourceInstanceArguments arguments = 1;
+    mz_storage_client.controller.ProtoCollectionMetadata storage_metadata = 2;
+    mz_repr.relation_and_scalar.ProtoRelationType typ = 3;
+}
+
+message ProtoSourceInstanceArguments {
+    optional mz_expr.linear.ProtoMapFilterProject operators = 1;
+}
