+++ src/compute/src/compute_state.rs
@@ -27,13 +27,12 @@ use timely::worker::Worker as TimelyWorker;
 use tokio::sync::{mpsc, Mutex};
 use uuid::Uuid;
 
-use mz_compute_client::command::{
-    ComputeCommand, ComputeCommandHistory, DataflowDescription, InstanceConfig, Peek,
-};
+use mz_compute_client::command::{ComputeCommand, ComputeCommandHistory, InstanceConfig, Peek};
 use mz_compute_client::logging::LoggingConfig;
 use mz_compute_client::metrics::ComputeMetrics;
 use mz_compute_client::plan::Plan;
 use mz_compute_client::response::{ComputeResponse, PeekResponse, SubscribeResponse};
+use mz_compute_client::types::dataflows::DataflowDescription;
 use mz_ore::cast::CastFrom;
 use mz_ore::tracing::OpenTelemetryContext;
 use mz_persist_client::cache::PersistClientCache;
