+++ src/compute/src/sink/subscribe.rs
@@ -21,7 +21,7 @@ use timely::progress::timestamp::Timestamp as TimelyTimestamp;
 use timely::progress::Antichain;
 
 use mz_compute_client::response::{SubscribeBatch, SubscribeResponse};
-use mz_compute_client::sinks::{ComputeSinkDesc, SinkAsOf, SubscribeSinkConnection};
+use mz_compute_client::types::sinks::{ComputeSinkDesc, SinkAsOf, SubscribeSinkConnection};
 use mz_repr::{Diff, GlobalId, Row, Timestamp};
 use mz_storage_client::controller::CollectionMetadata;
 use mz_storage_client::types::errors::DataflowError;
